requirements.txt
torch==2.0.1+cu118
transformers==4.47.0
datasets==4.0.0
scanpy==1.11.5
anndata==0.9.2
scvelo==0.3.2
cellrank==2.0.7
numpy==1.26.4
pandas==2.3.1
scipy==1.15.3
matplotlib==3.7.5
scikit-learn==1.1.3
pyyaml==6.0.3
tqdm==4.67.1
wandb==0.16.5
timm
einops
diffusers==0.30.1
einx
jax
celltypist
scib
